requirements.txt
# DrawingAI Pro - Dependencies
# Python 3.10+
# For reproducible installs on deployment targets, use requirements.lock.

# Azure OpenAI
openai>=1.12.0
msal>=1.26.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.0

# PDF Processing
pdfplumber>=0.9.0

# Image Processing & OCR
opencv-python>=4.8.0
Pillow>=10.0.0
pytesseract>=0.3.10

# Environment & Config
python-dotenv>=1.0.0

# HTTP
requests>=2.31.0

# GUI
customtkinter>=5.2.0

# Web UI
streamlit>=1.30.0
streamlit-autorefresh>=1.0.0

# Testing
pytest>=7.4.0
